+++ app/models/user.rb
@@ -1,6 +1,6 @@
 class User
   include Lotus::Entity
-  self.attributes = :email
+  self.attributes = :email, :username
 
   def auth_for(provider)
     AuthRepository.by_user_and_provider(id, provider)