+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 2014-07-03
 ----------
 
+* Store username
 * Dummy playlists with React.js
 
 2014-07-02